ref: long to int

+++ src/main/java/edu/agh/streamgraph/dynamicconnectedcomponents/EdgeSplitter.java
@@ -5,10 +5,10 @@ import org.apache.flink.types.NullValue;
 import org.apache.flink.util.Collector;
 import org.apache.flink.graph.Edge;
 
-public class EdgeSplitter implements FlatMapFunction<Edge<Long, NullValue> , ProcessMessage> {
+public class EdgeSplitter implements FlatMapFunction<Edge<Integer, NullValue> , ProcessMessage> {
     @Override
-    public void flatMap(Edge<Long, NullValue>  edge, Collector<ProcessMessage> collector) throws Exception {
+    public void flatMap(Edge<Integer, NullValue>  edge, Collector<ProcessMessage> collector) throws Exception {
         collector.collect(ProcessMessage.forEdgeIncoming(edge));
-        collector.collect(ProcessMessage.forEdgeIncoming(new Edge<Long, NullValue> (edge.f1, edge.f0, NullValue.getInstance())));
+        collector.collect(ProcessMessage.forEdgeIncoming(new Edge<Integer, NullValue> (edge.f1, edge.f0, NullValue.getInstance())));
     }
 }
